requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pypdf==4.0.0
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.1.0
pdf2image==1.16.3
img2pdf==0.5.1
docx2pdf==0.1.8